Update 1.2.1
- Bug fix: error on `isClaudeModel` in `Models/Anthropic.ts`
- Minor upgrade on how `ConversationManager` works. Now supporting caching in RAM only and both RAM and file system as options
- Corresponding updates to function calls involving `ConversationManager`

+++ src/ConversationManager.ts
@@ -4,11 +4,28 @@ import { v4 as uuid, validate as isUuid } from "uuid";
 
 export class ConversationManager {
 	protected cacheDir: fsSync.PathLike;
-	constructor(cacheDir: string) {
-		if (!cacheDir.endsWith(`/`)) throw new Error(`Cache directory error: dir not ended with \`/\``);
-		if (!fsSync.existsSync(cacheDir))
-			fsSync.mkdirSync(cacheDir, { recursive: true });
-		this.cacheDir = cacheDir;
+	protected cache: Map<string, Conversation>;
+	protected intervalTimer: NodeJS.Timer;
+
+	constructor(cacheToFile: boolean, options = {
+		/**
+		 * The directory in which the conversations are cached. Must end with `/`
+		 */
+		cacheDir: `data/conversations/`,
+		/**
+		 * The interval of which the conversation cache will be written into files. Number written in m
+		 */
+		saveInterval: 30 * 1000,
+	}) {
+		this.cache = new Map();
+
+		if (cacheToFile) {
+			if (!options.cacheDir.endsWith(`/`)) throw new Error(`Cache directory error: dir not ended with \`/\``);
+			if (!fsSync.existsSync(options.cacheDir))
+				fsSync.mkdirSync(options.cacheDir, { recursive: true });
+			this.cacheDir = options.cacheDir;
+			this.intervalTimer = setInterval(this.periodicUpdate.bind(this), options.saveInterval);
+		}
 	}
 
 	public async create(conversationMsg: ConversationMessage[]) {
@@ -20,7 +37,7 @@ export class ConversationManager {
 		return conversation;
 	}
 
-	public async getCopy(conversationId: string): Promise<Conversation> {
+	public async exportCopy(conversationId: string): Promise<Conversation> {
 		const conversation = await this.get(conversationId);
 		conversation.id = uuid();
 		await this.save(conversation);
@@ -28,12 +45,15 @@ export class ConversationManager {
 	}
 
 	protected async save(conversation: Conversation) {
-		await fsAsync.writeFile(`${this.cacheDir}${conversation.id}.json`,
-			JSON.stringify(conversation), { encoding: `utf-8` });
+		this.cache.set(conversation.id, conversation);
+		if (this.cacheDir)
+			await fsAsync.writeFile(`${this.cacheDir}${conversation.id}.json`,
+				JSON.stringify(conversation), { encoding: `utf-8` });
 	}
 
 	public has(conversationId: string) {
-		return fsSync.existsSync(`${this.cacheDir}${conversationId}.json`) && isUuid(conversationId);
+		return isUuid(conversationId) && (this.cache.has(conversationId) ||
+			((fsSync.existsSync(`${this.cacheDir}${conversationId}.json`) && this.cacheDir) || (!this.cacheDir)));
 	}
 
 	public async update(conversation: Conversation) {
@@ -44,15 +64,23 @@ export class ConversationManager {
 
 	public async get(conversationId: string): Promise<Conversation> {
 		if (!this.has(conversationId))
-			throw new Error(`Conversation Id does not exist: ${conversationId}`);
-		const data = await fsAsync.readFile(`${this.cacheDir}${conversationId}.json`, { encoding: `utf-8` });
-		return JSON.parse(data);
+			throw new Error(`[System] Conversation id ${conversationId} does not exist.`);
+		if (this.cache.has(conversationId))
+			return this.cache.get(conversationId) as Conversation;
+		else {
+			const data = await fsAsync.readFile(`${this.cacheDir}${conversationId}.json`, { encoding: `utf-8` });
+			return JSON.parse(data);
+		}
 	}
 
 	public async del(conversationId: string) {
 		if (!this.has(conversationId))
-			throw new Error(`Conversation Id does not exist: ${conversationId}`);
-		await fsAsync.unlink(`${this.cacheDir}${conversationId}.json`);
+			return null;
+		this.cache.delete(conversationId);
+		if (this.cacheDir) {
+			if (fsSync.existsSync(`${this.cacheDir}${conversationId}.json`))
+				await fsAsync.unlink(`${this.cacheDir}${conversationId}.json`);
+		}
 		return conversationId;
 	}
 
@@ -60,5 +88,11 @@ export class ConversationManager {
 		if (this.has(conversation.id))
 			return false;
 		await this.save(conversation);
+	}
+
+	protected periodicUpdate() {
+		this.cache.forEach(async (val, key) => {
+			await fsAsync.writeFile(`${this.cacheDir}${key}.json`, JSON.stringify(val), { encoding: `utf-8` });
+		});
 	}
 }
